Add configurable virtualenv python discovery
SublimePythonIDE will now try to automatically discover the python interpreter
in the project's virtualenv. By default, it will search for a directory named
'venv'. This is configurable with the `virtualenv_dir_name` setting.

If no virtualenv interpreter is discovered, it will fall back on the system
interpreter.

The `python_interpreter` setting still takes precedence over the virtualenv
interpreter discovery. Is this what we want ?